Reduce the config file to the bare minimum and make it optional
Now Ubuntu is an option in the code itself, there is no need to specify
it's commands in the config file any more. Thus we can reduce it to just
a container_name, the image to use and the package manager to use for
that image.

The config file is now also completely optional. If it isn't configured
(either by a distribution package or by the user itself) apx will try to
detect the config values from the host system. If it's an unsupported
setup (e.g. no known package manager) then it'll error out and instruct
the user to setup a config file.