Move distro definition to a struct and change pkgmanager flags for distro flags
Distro definition is now a single struct with all relevant info for that
distribution in a single place. When new distributions are now added as
built-in sources, they only have to be specified in core/distros.go and
as a commandline flag.

This also makes it possible to use more containers than just the
built-in supported ones, as long as the package manager used in that
container is a supported one. E.g. a Debian container can now be used
by specifying it in the config file.

Aside from that the package manager flags for commands are replaced for
distribution flags. For the end user the package manager used internally
isn't interesting, the source of the installed package however is. So
--ubuntu is more relevant than --apt, which could mean either Ubuntu,
Debian, or some other apt using distro.